FIXES for exclude app from update PR
