+++ bench/commands/update.py
@@ -6,11 +6,8 @@ from bench.utils import (update_bench, validate_upgrade, pre_upgrade, post_upgra
 	update_requirements, update_node_packages, backup_all_sites, patch_sites, build_assets, restart_supervisor_processes)
 from bench import patches
 
-# TODO: Not DRY
-
 
 @click.command('update')
-@click.option('--exclude-apps', is_flag=True, help="Exclude pull for the apps")
 @click.option('--pull', is_flag=True, help="Pull changes in all the apps in bench")
 @click.option('--patch', is_flag=True, help="Run migrations for all sites in the bench")
 @click.option('--build', is_flag=True, help="Build JS and CSS artifacts for the bench")
@@ -21,10 +18,12 @@ from bench import patches
 @click.option('--no-backup', is_flag=True)
 @click.option('--force', is_flag=True)
 @click.option('--reset', is_flag=True, help="Hard resets git branch's to their new states overriding any changes and overriding rebase on pull")
-def update(exclude_apps=False, pull=False, patch=False, build=False, bench=False, auto=False, restart_supervisor=False, requirements=False, no_backup=False, force=False, reset=False):
-	"""Update bench"""
+def update(pull=False, patch=False, build=False, bench=False, auto=False, restart_supervisor=False, requirements=False, no_backup=False, force=False, reset=False):
+	"Update bench"
+
 	if not (pull or patch or build or bench or requirements):
 		pull, patch, build, bench, requirements = True, True, True, True, True
+
 	if auto:
 		sys.exit(1)
 
@@ -34,7 +33,6 @@ def update(exclude_apps=False, pull=False, patch=False, build=False, bench=False
 	if bench and conf.get('update_bench_on_update'):
 		update_bench()
 		restart_update({
-				'exclude-apps': exclude_apps,
 				'pull': pull,
 				'patch': patch,
 				'build': build,
@@ -57,9 +55,9 @@ def update(exclude_apps=False, pull=False, patch=False, build=False, bench=False
 		print("This would take significant time to migrate and might break custom apps.")
 		click.confirm('Do you want to continue?', abort=True)
 
-	_update(exclude_apps, pull, patch, build, bench, auto, restart_supervisor, requirements, no_backup, force=force, reset=reset)
+	_update(pull, patch, build, bench, auto, restart_supervisor, requirements, no_backup, force=force, reset=reset)
 
-def _update(exclude=False, pull=False, patch=False, build=False, update_bench=False, auto=False, restart_supervisor=False,
+def _update(pull=False, patch=False, build=False, update_bench=False, auto=False, restart_supervisor=False,
 		requirements=False, no_backup=False, bench_path='.', force=False, reset=False):
 	conf = get_config(bench_path=bench_path)
 	version_upgrade = is_version_upgrade(bench_path=bench_path)
@@ -70,7 +68,7 @@ def _update(exclude=False, pull=False, patch=False, build=False, update_bench=Fa
 	before_update(bench_path=bench_path, requirements=requirements)
 
 	if pull:
-		pull_all_apps(bench_path=bench_path, reset=reset, exclude=exclude)
+		pull_all_apps(bench_path=bench_path, reset=reset)
 
 	if requirements:
 		update_requirements(bench_path=bench_path)
