+++ bench/utils.py
@@ -278,6 +278,7 @@ def update_bench():
 
 	# bench-repo folder
 	cwd = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+
 	exec_cmd("git pull", cwd=cwd)
 
 def setup_sudoers(user):
