+++ patches.txt
@@ -1,0 +1,1 @@
+bench.patches.v3.deprecate_old_config
